requirements.txt
fastapi
uvicorn[standard]
torch==2.6.0
torchvision
transformers==4.39.3
Pillow
python-multipart
tqdm
matplotlib
seaborn
torchinfo
numpy
requests
